SC-97: disallow compiling functions to V1/V2 contexts
